multi-agent-system/requirements.txt
# Core dependencies
pydantic-ai>=0.0.1
fastmcp>=0.1.0
pyyaml>=6.0
jinja2>=3.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# Langfuse/Telemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp-proto-http>=1.20.0

# HTTP server
starlette>=0.27.0
uvicorn>=0.23.0
fastapi>=0.104.0

